Release : 220407 V0.9.0

+++ src/main/java/com/ssafy/tnt/service/NewsCrawlerService.java
@@ -10,7 +10,7 @@ import com.ssafy.tnt.model.NewsDTO;
 
 public interface NewsCrawlerService {
 	List<NewsDTO> crawlNews() throws Exception;
-	int komoran() throws IOException;
+	double komoran() throws IOException;
 	void getTF() throws IOException;
 	void insertKeyWord() throws IOException;
 	void TFIDF(HashMap<String, Double> tmpTFMap) throws IOException;
